+++ src/ui/index.ts
@@ -27,4 +27,3 @@ export * from './TextArea/TextArea';
 export * from './TextField/TextField';
 export * from './ToggleButton';
 export * from './Tooltip/Tooltip';
-export * from './Toast/Toast';
